pyproject.toml
[project]
name = "rag-leis-juridicas"
version = "0.1.0"
description = "Projeto que visa dar informações sobre leis tributárias via rag"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.3.8",
    "fastapi>=0.124.2",
    "firecrawl-py>=4.10.3",
    "openai>=2.9.0",
    "pydantic-settings>=2.12.0",
    "qdrant-client>=1.16.1",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]

[tool.setuptools.packages.find]
include = [
    "agents*",
    "config*",
    "prompts*",
    "schemas*",
    "vectordb*",
    "scripts*",
]
